refactor: move includes to implementation (partial) Move all possible platform-specific and internal-use includes to the compiled implementation unit.
Not all platform includes have been moved at this time, as the
`UIMenu` struct definition utilizes these on MacOS/Cocoa and
Essence. This will be addressed in a future patch.

Also, within the implementation section, the platform includes
are not currently within each platform section (although they are
guarded appropriately). This is due to the global `ui` struct
utilizing them. This will be addressed in a future patch.

| Like prior patches: This improves separation of concerns,
| enhances code structure, and is part of work to move
| platform-specific and internal functionality and headers
| to the compiled implementation unit.
|
| Further work to be done.