build: improve user-friendliness of the Makefile Add functionality to only build particular configurations of the examples, for example, most users will want to run commands like:
 $ make base
 $ make debug
 $ make freetype
 $ make freetype_unicode
 $ make sse2

These will build all examples only for the specified variant. See
`CONFIGS` in the Makefile for all available variants.

All variants will be built with `make all`, which is used by the CI
pipeline for testing changes to the library.

Also, provide nicer output when running `make clean`, and correct a
couple of typos.